traffic_signs_Vision_LLM/requirements.txt
openai>=1.30.0
streamlit
Pillow
python-dotenv